feat(ux): dedicated "Canvas" tab in /settings (#1359) (#1424)
Canvas toolbar toggles (Show Test Trigger, Show Dry Run) lived under
"Appearance" which is theme + accent colour — wrong home. Moves them to
a new "Canvas" tab in workspace /settings, adjacent to Appearance.

- **New apps/web/src/components/settings/CanvasPanel.tsx** — extracted
  from PreferencesPanel. Same workspace-scoped storage via
  usePreferences() (show_test_trigger / show_dry_run). Local ToggleRow
  helper.
- **apps/web/src/app/(app)/settings/page.tsx** — adds canvas tab
  between Appearance and Proxy in TABS + panels dict.
- **apps/web/src/components/settings/PreferencesPanel.tsx** — strip the
  Canvas options section, the ToggleRow helper (no longer used), and
  the usePreferences import. Net -53 lines.

No behaviour change: same toggles, same storage, same effect on the
workflow canvas — just found where users expect.

Ref #1359.
