+++ apps/web/src/app/(app)/settings/page.tsx
@@ -6,6 +6,7 @@ import { useAuthFetch } from "@/hooks/useAuthFetch"
 import { API } from "@/lib/api"
 import { SettingsShell, type SettingsTab } from "@/components/SettingsShell"
 import { useWorkspace } from "@/lib/WorkspaceContext"
+import CanvasPanel from "@/components/settings/CanvasPanel"
 import EnvironmentsManager from "@/components/settings/EnvironmentsManager"
 import MembersManager from "@/components/settings/MembersManager"
 import PreferencesPanel from "@/components/settings/PreferencesPanel"
@@ -13,12 +14,13 @@ import ProxySettings from "@/components/settings/ProxySettings"
 import LLMPrimitivesPanel from "@/components/settings/LLMPrimitivesPanel"
 import RateLimitsPanel from "@/components/settings/RateLimitsPanel"
 
-type Tab = "credentials" | "llm_primitives" | "members" | "preferences" | "proxy" | "rate_limits"
+type Tab = "credentials" | "llm_primitives" | "members" | "preferences" | "canvas" | "proxy" | "rate_limits"
 
 const TABS: readonly SettingsTab<Tab>[] = [
   { key: "credentials",    label: "Vault" },
   { key: "llm_primitives", label: "LLM Model Primitives" },
   { key: "preferences",    label: "Appearance" },
+  { key: "canvas",         label: "Canvas" },
   { key: "proxy",          label: "Proxy" },
   { key: "rate_limits",    label: "Rate limits", adminOnly: true },
   { key: "members",        label: "Members & roles", adminOnly: true },
@@ -184,6 +186,7 @@ function SettingsPageInner({ isAdmin, workspaceId, getToken }: { isAdmin: boolea
     credentials:    <EnvironmentsManager isAdmin={isAdmin} />,
     llm_primitives: <LLMPrimitivesPanel workspaceId={workspaceId} isAdmin={isAdmin} />,
     preferences:    <PreferencesPanel />,
+    canvas:         <CanvasPanel />,
     proxy:          <ProxySettings workspaceId={workspaceId} getToken={getToken} />,
     rate_limits:    <RateLimitsPanel isAdmin={isAdmin} />,
     members:        <MembersManager />,
